refactor(intruder_alarm::doubly): renames `doubly` to `list`.
